[21/n] reconfigurator-cli] add tuf-assemble command for TUF repo generation (#8469)
I need the ability to generate TUF repos for upcoming tests, and
adapting the `target-release.txt` test seemed like the easiest way to do
it.

This also means that we no longer need to handle that test through a
custom wrapper -- the standard `test-scripts.rs` test harness works well
for it now.

This is a standalone PR that can land independently of the rest.