Bump OPTE to v0.37.386
Brings in the performance work which gets us out of promiscuous mode,
and then builds on having access to multiple packets at a time in the Rx
callback.

Tested on racklettes on a few occasions up til now, where we're looking
at >10Gbps single TCP stream (for LRO-aware guests). Helios-2.8 image is
able to load and interact with the driver, so we're good on an a4x2
front.

0.37.387 exists, but requires rustc v1.88.0+ due to if-let chains.