[22/n] [reconfigurator-cli] specify sled by serial number (#8489)
These serial numbers are somewhat easier to parse in many cases, and
also synergize with the next commit in this stack (which allows setting
a specific policy on each sled while loading the initial example).

Switch one of the tests over as an example -- I hope its intent becomes
clearer.