sled-agent-config-reconciler: actually call `on_time_sync()` (#8509)
@askfongjojo noticed that our `uptime`s on dogfood are all nonsense:

```
 8  BRM44220011        ok: 14:05:34    up 14066 day(s), 14:06,  1 user,  load average: 5.54, 5.60, 5.62
 9  BRM44220005        ok: 14:05:34    up 14066 day(s), 14:06,  1 user,  load average: 17.51, 17.81, 17.81
10  BRM42220009        ok: 14:05:35    up 14066 day(s), 14:06,  1 user,  load average: 15.00, 14.46, 14.01
11  BRM42220006        ok: 14:05:35    up 14066 day(s), 14:06,  0 users,  load average: 11.49, 11.21, 11.10
12  BRM42220057        ok: 14:05:35    up 14066 day(s), 14:06,  0 users,  load average: 2.88, 2.62, 2.03
...
```

I _think_ #8064 introduced this. It shuffled around how time sync is
checked and added a callback that the config-reconciler is supposed to
run when it detects time has synchronized; that callback is responsible
for rewriting `uptime` (among other things), but it never actually
executes the callback. This PR fixes that.

However, we have some racklettes that are running commits that include
#8064 that have reasonable uptimes. I'm not sure how that's possible -
is there some other way `uptime` can be correct if sled-agent doesn't
fix it?