Planner/builder split: Move image source decisions to the planner (#8472)
* `BlueprintBuilder` no longer decides on image sources for zones. All
the `sled_add_zone_*` functions now take an extra `image_source`
argument.
* `is_zone_ready_for_update()` is now a method in the planner, not the
builder. This is used both when deciding whether a zone can be updated
and also when choosing the image source for new zones to be added.
* `can_zone_be_shut_down_safely()` is now a separate method in the
planner. This is only used when deciding whether a zone can be updated.
* `zone_image_source` is now a method on `TargetReleaseDescription`. It
returns an error if we have a TUF repo that doesn't have exactly one
matching artifact for a given `ZoneKind`. In the planner, this bubbles
out a few ways:
* we'll fail planning if we try to add a zone that isn't in the target
release TUF repo
* we'll refuse to update a zone if it isn't in the target release TUF
repo (but this won't fail planning)
* we'll refuse to update Nexus (because it won't be able to tell whether
whatever zone type is missing has itself been updated)