[sled-agent] Read boot partition of M.2 drives (PR 1/2) (#8450)
This PR is basically just adding the `host_phase_2` module to
`sled-agent-config-reconciler`, which reads the boot partition of the
two M.2 slots (using multiple-of-the-disk-block-size reads, which are
required when accessing the raw device). The results of these reads is
only logged as of this PR; the followup adds it to inventory.

There are a handful of other cleanup changes that snuck in here due to
dropping an out-of-date `#![allow(dead_code)]`.