[sled-agent] Don't bounce zones if their config changes only by an identical `image_source` (#8514)
Prior to this PR, any change to the `OmicronZoneConfig` of an
already-running zone resulted in that zone being shut down and
restarted. With this PR, we allow a single kind of change that does not
restart the zone. If the `image_source` swaps from `InstallDataset` to
`Artifact { hash }` (or vice versa) _and_ the hash of this zone in the
install dataset exactly matches the `Artifact { hash }`, we don't need
to do anything: we're already running the exact zone as desired, just
started from a different place.

Fixes #8463.

Makes #8510 slightly worse; yet another `ZoneKind::make_a_string()`
method. (This one is at least guaranteed-by-test to be an extension of
an existing one. I don't feel a lot better about that though.)

I'll test this on a racklette before merging; will put notes below once
I do. But I think this is contained enough that it can be reviewed
before that's done.