Update Propolis
Changes here are:
* update zerocopy to 0.8.25 (Propolis#910)
* Rust 1.87.0 clippy lints (Propolis#911)
* Support guest memory sizes >=1028 GiB (Propolis#912)
* Fix new lints (Propolis#914)
* server: build Tokio runtime using `oxide-tokio-rt` (Propolis#913)
* Handle CPUID subleaves more precisely when using Bhyve defaults (Propolis#915)
* update tokio-dtrace to 0.1.1 (and tokio to 1.46.1) (#916)

Since the Propolis update increases the max viable memory size, I'm
bumping that to 1.5 TiB here as well (it's the largest size I've tested
on `dublin`)

The CPUID changes in #915 are guest-visible and on production racks will
have us stop advertising platform QoS features, whose MSRs aren't made
available to guests anyway.

The tokio-dtrace change is internal to the Propolis artifacts here, so
we still need a (separate!) bump of tokio-dtrace to get that update in
Omicron proper for sled-agent, Nexus, and other binaries here.