[24/n] [reconfigurator-planning] support no-op image source updates (#8486)
RFD 556 discusses part of the mupdate recovery process being a zone
image source switch from the install dataset to the TUF repo depot.

During discussions, we figured out that this doesn't have to be tied to
the mupdate override stuff for the most part, and that this is a process
we are allowed to perform on any zones where a mupdate override isn't in
place.

This depends on some pending work in Sled Agent to treat image source
switches as no-ops if the underlying hash is the same.