[sled-agent] Report boot partition contents of M.2 drives in inventory (PR 2/2) (#8451)
Add a `BootPartitionContents` structure to the inventory reported by the sled-agent-config-reconciler which contains:

* which slot we booted from (A or B)
* the contents of each slot

each of which is a `Result` in case we failed to determine any of those
three items. In practice, I'd expect us to basically never fail to
report which slot we booted from, and to only fail on the contents of a
slot if that slot doesn't have a valid phase 2 image in it.