Delete `builtin_roles` (#7477)
This came up in a meeting with @davepacheco — storing the roles in the
DB and letting them be listed through the API is a legacy idea with no
practical application that we know of. When I heard that I was curious
how much could be deleted.

Fundamentally, all this machinery I've deleted was just in service of
the `role_view` and `role_list` endpoints, which let you get something
like this list:

```
fleet     admin
fleet     collaborator
fleet     viewer
silo      admin
silo      collaborator
silo      viewer
project   admin
project   collaborator
project   viewer
```

I think we thought this would be a dynamic list of roles, but all these
things are hard-coded. As far as I can tell there is no relationship to
the actual authorization system — this is basically just an unrelated
copy of that stuff. This claim borne out by the fact that deleting all
this stuff doesn't affect anything else, i.e., doesn't cause any tests
to fail.