properly record TUF artifacts we already have (#8571)
Two TUF repositories with different system versions can contain the same
artifacts (e.g. Hubris archives, which have their own release cadence).
When inserting a `TufRepoDescription`, we generate new IDs for all
artifacts regardless of whether they already exist in the database; we
then look for duplicates and avoid inserting already-present artifacts.
But when those artifacts are associated with the repository in the
`tuf_repo_artifact` table, we incorrectly use the ID generated for this
upload and not the ID of the artifact in the database. This results in
duplicate artifacts being "missing" from the repo description when
pulled from the database.