Represent resolved firewall `filter_hosts` as a `HashSet` (#8315)
Previously, we have been adding one `filter_host` entry for *every NIC*
in the VPC when using a `VpcFirewallRuleHostFilter::Vpc(...)`. This PR
fixes this in two ways:

* We no longer linear-scan the NIC list to attempt to resolve
`vpc_name`->`Vni` mappings per rule. We build a `HashMap` of these early
on.
* This isn't *practically* a problem given that we're only working in
terms of a single VPC today, but the principle is correct for when we do
support e.g. VPC peering.
* `filter_hosts` is now encoded as a `HashSet` in nexus and sled-agent,
which ensures this property holds even under users' attempts to
introduce duplicate host specifiers.

Closes #8305.