Port/ICMP parameter ranges should enforce `first <= last` (#8578)
The console correctly prevents ranges like `3000-1000` for
`L4PortRange`s, but we don't catch this in the API itself. This PR
enforces that on the frontend, with some allowances for the fact that we
might still have such rows in CRDB.

I was torn between enforcing this using transparent types to apply the
first/last constraint, which worked in terms of being invisible in the
JsonSchema, versus what I did here which is just checking the invariant
on conversion to a db::VpcFirewallRule. Unfortunately the former was far
messier, and ended up splitting the logic such that it was a little
harder to intuit what was checked and where. So, we validate this next
to the filter array length checks, which felt fairly natural.