[mgs] Don't drop the ereport socket
Turns out that dropping this cancels the receive handler task for the
ereport socket, which results in ereport collection being totally
busted. This wasn't caught until I actually tested the system end-to-end
on a racklette with a real SP emitting ereports, since the omicron
integration tests don't actually use a single shared socket, and
`faux-mgs` just puts it on the stack, so none of this was actually
exercised outside of a production-like environment. Cool.

Shoutout to @mkeeter for finding the bug!