TQ: Encrypt all necessary prior rack secrets
This is an update to the trust quorum protocol that no longer relies on
tracking the last committed configuration. Instead it encrypts multiple
rack secrets to allow upgrading past only one missed reconfiguration if
a sled is offline.

This is based of the TLA+ spec and what is written in RFD 238.