Add desired host phase 2 contents to blueprint (#8570)
This is a followup to the stack of PRs that added desired host phase 2
contents to `OmicronSledConfig` (#8538 and its followups). We now have a
`BlueprintHostPhase2DesiredSlots` inside each `BlueprintSledConfig` that
specifies the desired contents (artifact by hash-and-version or "just
the current contents", analogous to the install dataset for zones) for
each of the two phase 2 slots.