[29/n] [reconfigurator] separate out noop image source decision making (#8596)
While working on #8572, we realized that we need to put in a condition
saying that if any zones can't be switched over to Artifact, we must not
clear the mupdate override field. This effectively requires information
about eligibility to be available in two spots: while updating the
mupdate override, and while doing these noop conversions.

In order to do that more easily, this PR builds up a decision tree by
sled.