Only query available SPs for task dumps
The MGS `sp_all_ids` method will return all SPs that could potentially
be present, which resulted in noise in the Nexus logs from attempts to
connect to SPs in empty cubbies.

Switch to the `ignition_list` method, which allows us to filter for SPs
that are present, powered on, and have not faulted.