[30/n] [omdb] move inventory collection display to nexus-types (#8613)
There are no major code changes in this PR -- it's pure movement +
replacing `println!` calls with `writeln!`.

There's one minor change, which is that the warning about errors now
goes to stdout. I think it's okay.

Tests are coming in the next PR as part of adding support to
reconfigurator-cli.

Part of #8611.