[31/n] [reconfigurator-cli] add inventory-show command (#8617)
Use the display logic implemented in #8613 to show inventory collections
in reconfigurator-cli.

I had to change some of the timestamps to be in RFC 3339 format to allow
redactions to work properly.

Kind of addresses #8611, in the sense that we now have tests that render
collection output and ensure it looks reasonable.