More plumbing, rethink Probe firewall rule insertion.
This should wire up any remaining compile issues around the extra IDs
we're inserting for rules/routes/NICs.

The bigger change that needed to be made is around how firewall rules
make their way to probes. Today we start by allowing all traffic, and
then, assumedly, waiting until a change is made to the VPC firewall
rules to propagate the true routeset to the probe. However, that doesn't
fit cleanly in an obvious way with the reconciler pattern around
`ProbeInfo`.

What we do now is to repurpose the (now-unused)
`sled_firewall_rules_request` internal endpoint to accept a VPC, rather
than implicitly working only for services. Services themselves have long
since moved onto relying on the `"service_firewall_rule_propagation"` task.
When any probes are added, we now request an update for the VPC each
resides within. Ideally, these *should* be part of the probe on creation
(though this is arguably less crucial than it is for instances).