TQ: Add `NodeCtx` for use in `Node` API (#8629)
Rather than passing in separate parameters to `Node` methods, we pass in
a context shared with the caller. Importantly, due to the use of wrapper
traits, the caller and methods are only able to mutate things that they
each should be able to mutate.

This makes the function signatures shorter and also eliminates the case
where we end up returning an `Option<PersistentState>`. This was getting
tedious, not just here, but in other code not yet pushed. Since the
caller already has to check if there are any outgoing messages to send
after each api call, they can also check to see if the persistent state
has changed. This isn't much of a burden and simplifies the `Node`
internals somewhat. It also eliminates a clone. We can add other state
to the context as needed over time.