helios/deploy: Wait for inventory collection (#8650)
This adds another "wait for a thing to happen" loop in between waiting
for the RSS -> Nexus handoff and waiting for all the zpools to show up.
Attempting to query for the zpools fails if there hasn't been an
inventory collection yet, so prior to this change we were racing.

Fixes #8647.

---------

Co-authored-by: iliana etaoin <iliana@oxide.computer>