[36/n] [nexus-types] make PendingMgsUpdates use iddqd::IdOrdMap (#8655)
This enables usage of `&BaseboardId` in upcoming work that needs access
to the Entry API on `PendingMgsUpdates`.

I had to make a few changes to the diff code for what appear to be
variance reasons -- I'm not entirely sure why because r-a reports that
the lifetime was invariant both before and after. The changes are
relatively minor, though.