Do not panic in switch synchronization task
There were some lingering `expect()` and `unwrap()` statements
in the sync switch configuration background task. Some of these
were in places we believed to be impossible to hit, but a recent
core dump shows that we have somehow triggered one of them, so
we're now logging an error and continuing in each case.