[39/n] prepare sled-agent reconciler for honoring mupdate overrides (#8688)
This PR prepares the sled-agent config reconciler to honor mupdate
overrides. In particular, the logic that requires bouncing zones needs
to be updated to consider zone image locations after mupdate overrides
have been considered, not before.

Doing this required some refactoring. The biggest one is that looking up
zone image sources is no longer done through the image resolver and/or
within sled-agent's `services.rs`, but rather by gathering and querying
the resolver status within the config reconciler. This is a nice
improvement overall, particularly because it means we grab the lock once
per reconciler run rather than each time we need to look up a zone's
image source.

We do not actually honor the mupdate override yet, though all the pieces
are now in place. We'll combine the PRs to honor the override and update
the blueprint logic into one, within #8456.