Add basic TUF repo check after generation (#8722)
There are an extended number of tests designed to catch issues with TUF
repos. Despite this, it is still possible (though unlikely) to generate
a TUF repo that fails one or more checks. Nobody likes broken builds so
add an extra last check to verify that we can do "something" with the
TUF repo.