[sp-sim] Enhancements to support host OS testing (#8731)
This is pulled out of the upcoming Reconfigurator host OS update work to
make that PR slightly smaller. This has a few small enhancements to our
SP simulator:

* Simulated sleds now always report something for both phase 1 slots.
The initial values for the two slots are different. (This isn't really
necessary but makes debugging some tests easier, since without this both
slots report the same hash value for their contents.)
* We expose a count of the number of times we've performed a power state
transition. (Will be used to test whether an update triggers a reboot.)
* Implement `component_get_active_slot()` for host boot flash.
* Simulated sled only: We expose a watch channel with the current power
state, and if the state is A0, it also includes which host boot flash
slot was active at the time we transitioned to A0. (We'll hook this up
to a fake sled-agent so it can report a different boot disk after an
sp-sim "reboot".)