nexus-mgs-updates: Cleanup and refactoring to support host OS updates (#8736)
This is mostly a "refactoring things a bit" PR to reduce diff churn with
the upcoming host OS updates PR, but there are some small behavioral
changes, mostly but not entirely affecting tests:

* In tests, simulated sled-agent now reports valid boot partition
contents instead of an error
* Fix a small bug in calculating elapsed time that caused
`reconfigurator-sp-updater` to always show "updating for 0s" instead of
the actual elapsed time
* Replace `RESET_TIMEOUT` with a function so we can have different
values for different kinds of updates. (This is mostly important because
the "reset a host" timeout needs to be _much_ longer than "reset an
SP/RoT" timeouts, but also allows us to go back to 60s for sled/psc SPs
while keeping 120s for switch SPs.)
* In `nexus-mgs-updates` tests, create MGS clients that have longer
timeouts. (Reason for this is in comments; this definitely affects the
host OS tests that don't exist as of this PR, but other tests that are
on main are already racy and can fail the same way on a busier system.)
* In `nexus-mgs-updates` tests, make the `deployed_caboose` field an
`Option`, because host OS artifacts won't have one. I don't love this in
that it's an option dependent on other fields (i.e., SP/RoT/RoT
bootloader updates always fill this with `Some(_)`, and host OS updates
always have `None`). Usually I'd use an enum to squash the impossible
states out, but (a) this is test-only code and (b) using that enum in
tests seems slightly more onerous than just unwrapping. Happy to change
this if others feel differently though.