[Reconfigurator] Don't reevaluate impossible MGS updates for a few minutes (#8760)
This PR adds `ignore_impossible_mgs_updates_since` to `PlanningInput`,
and changes the planner to keep rather than reevaluate MGS updates that
it deems impossible if they are newer than
`ignore_impossible_mgs_updates_since`. This avoids churn when the act of
updating itself renders the planned update (temporarily) impossible. See
#8483 for context.

Fixes #8483.

Also fixes #8484 (not directly relevant to the main changes, but is just
a couple new lines in `do_plan_mgs_updates()`).

I'll test this on a racklette and post results in a comment before
merging.